backend/foodgram/requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==22.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.17
djangorestframework==3.15.2
django-templated-mail==1.1.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.1.1
flake8==6.0.0
flake8-isort==6.0.0
idna==3.6
iniconfig==1.1.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
numpy==2.0.0
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==21.3
Pillow==9.0.0
pluggy==0.13.1
prettytable==3.7.0
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2022.6
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.5.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.2.6
webcolors==1.11.1
django-filter==23.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.3 
pyshorteners
